requirements.txt
fastapi==0.116.0
slowapi==0.1.9
fastapi-pagination==0.14.3
python-multipart==0.0.5
sqlalchemy==1.4.46
sqlalchemy-utils==0.39.0
pydantic==2.12.0
psycopg2-binary==2.9.10
python-jose==3.3.0
cryptography==39.0.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
uvicorn[standard]==0.34.0
jinja2==3.1.2
Pillow==11.0.0
discord-webhook==1.3.1
pyotp==2.9.0
qrcode[pil]==7.4.2
